refactor out log collect into it's own package for easier testing
